libxml++@3: rename from `libxml++3`
Also:
* update homepage
* remove unnecessary `ENV.cxx11`
* update `meson` commands to current style
* simplify test with `pkg-config`
