+++ Formula/lib/libxml++@3.rb
@@ -1,0 +1,43 @@
+class LibxmlxxAT3 < Formula
+  desc "C++ wrapper for libxml"
+  homepage "https://libxmlplusplus.github.io/libxmlplusplus/"
+  url "https://download.gnome.org/sources/libxml++/3.2/libxml++-3.2.4.tar.xz"
+  sha256 "ba53f5eaca45b79f4ec1b3b28bc8136fce26873cd38f2e381d9355289e432405"
+  license "LGPL-2.1-or-later"
+
+  livecheck do
+    url :stable
+    regex(/libxml\+\+[._-]v?(3\.([0-8]\d*?)?[02468](?:\.\d+)*?)\.t/i)
+  end
+
+  depends_on "meson" => :build
+  depends_on "ninja" => :build
+  depends_on "pkg-config" => [:build, :test]
+  depends_on "glibmm@2.66"
+
+  uses_from_macos "libxml2"
+
+  def install
+    system "meson", "setup", "build", *std_meson_args
+    system "meson", "compile", "-C", "build", "--verbose"
+    system "meson", "install", "-C", "build"
+  end
+
+  test do
+    (testpath/"test.cpp").write <<~EOS
+      #include <libxml++/libxml++.h>
+
+      int main(int argc, char *argv[])
+      {
+         xmlpp::Document document;
+         document.set_internal_subset("homebrew", "", "https://www.brew.sh/xml/test.dtd");
+         xmlpp::Element *rootnode = document.create_root_node("homebrew");
+         return 0;
+      }
+    EOS
+    command = "#{Formula["pkg-config"].opt_bin}/pkg-config --cflags --libs libxml++-3.0"
+    flags = shell_output(command).strip.split
+    system ENV.cxx, "-std=c++11", "test.cpp", "-o", "test", *flags
+    system "./test"
+  end
+end
