+++ formula_renames.json
@@ -112,6 +112,7 @@
   "libpython-tabulate": "python-tabulate",
   "libsasl2": "cyrus-sasl",
   "libtorch": "pytorch",
+  "libxml++3": "libxml++@3",
   "linux-headers": "linux-headers@4.4",
   "llama": "walk",
   "lua51": "lua@5.1",
